requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
pyyaml>=6.0
tqdm>=4.65.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
git+https://github.com/openai/CLIP.git